--- FacebookWinFormsApp.sln
@@ -1,31 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.31205.134
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "FacebookWinFormsApp", "FacebookWinFormsApp\FacebookWinFormsApp.csproj", "{905D5BF1-A7E4-4440-85E8-84772F5312D5}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Debug|x64 = Debug|x64
-		Release|Any CPU = Release|Any CPU
-		Release|x64 = Release|x64
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{905D5BF1-A7E4-4440-85E8-84772F5312D5}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{905D5BF1-A7E4-4440-85E8-84772F5312D5}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{905D5BF1-A7E4-4440-85E8-84772F5312D5}.Debug|x64.ActiveCfg = Debug|x64
-		{905D5BF1-A7E4-4440-85E8-84772F5312D5}.Debug|x64.Build.0 = Debug|x64
-		{905D5BF1-A7E4-4440-85E8-84772F5312D5}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{905D5BF1-A7E4-4440-85E8-84772F5312D5}.Release|Any CPU.Build.0 = Release|Any CPU
-		{905D5BF1-A7E4-4440-85E8-84772F5312D5}.Release|x64.ActiveCfg = Release|x64
-		{905D5BF1-A7E4-4440-85E8-84772F5312D5}.Release|x64.Build.0 = Release|x64
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {5D8BC6CD-79F1-4070-9346-45BF5E5E6568}
-	EndGlobalSection
-EndGlobal
