--- FacebookWinFormsApp/packages.config
@@ -1,7 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="cef.redist.x64" version="87.1.13" targetFramework="net472" />
-  <package id="cef.redist.x86" version="87.1.13" targetFramework="net472" />
-  <package id="CefSharp.Common" version="87.1.132" targetFramework="net472" />
-  <package id="CefSharp.WinForms" version="87.1.132" targetFramework="net472" />
-</packages>
